Define dlong to long long instead of int64_t.

+++ calma/CalmaWrite.c
@@ -21,7 +21,6 @@ static char rcsid[] __attribute__ ((unused)) ="$Header: /usr/cvsroot/magic-8.0/c
 #endif  /* not lint */
 
 #include <stdio.h>
-#include <stdint.h>
 #include <string.h>
 #include <ctype.h>
 #include <sys/types.h>
@@ -2401,7 +2400,7 @@ calmaOutR8(d, f)
     int c, i, sign, expon;
 
     /* mantissa must be 64 bits for this routine to work correctly */
-    uint64_t mantissa;
+    unsigned long long mantissa;
 
     mantissa = 0;
     if (d == 0.0)
